Squash commits

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/IexQuote.java
@@ -42,9 +42,9 @@ public class IexQuote {
     private Double iexMarketPercent;
     private Integer iexVolume;
     private Integer avgTotalVolume;
-    private Integer iexBidPrice;
+    private Double iexBidPrice;
     private Integer iexBidSize;
-    private Integer iexAskPrice;
+    private Double iexAskPrice;
     private Integer iexAskSize;
     private Double iexOpen;
     private Long iexOpenTime;
@@ -379,11 +379,11 @@ public class IexQuote {
       this.avgTotalVolume = avgTotalVolume;
     }
 
-    public Integer getIexBidPrice() {
+    public Double getIexBidPrice() {
       return iexBidPrice;
     }
 
-    public void setIexBidPrice(Integer iexBidPrice) {
+    public void setIexBidPrice(Double iexBidPrice) {
       this.iexBidPrice = iexBidPrice;
     }
 
@@ -395,11 +395,11 @@ public class IexQuote {
       this.iexBidSize = iexBidSize;
     }
 
-    public Integer getIexAskPrice() {
+    public Double getIexAskPrice() {
       return iexAskPrice;
     }
 
-    public void setIexAskPrice(Integer iexAskPrice) {
+    public void setIexAskPrice(Double iexAskPrice) {
       this.iexAskPrice = iexAskPrice;
     }
